build(fields): renamed ControlledJSONField to ControlledJsonField.
